+++ handlers/__init__.py
@@ -1,0 +1,7 @@
+from handler import Handler
+from singup import Singup
+from welcome import Welcome
+from login import Login
+from logout import Logout
+from newpost import NewPost
+
