+++ index.yaml
@@ -9,6 +9,7 @@ indexes:
 # manually, move them above the marker line.  The index.yaml file is
 # automatically uploaded to the admin console when you next deploy
 # your application using appcfg.py.
+
 - kind: BlogPost
   properties:
   - name: created_by
