+++ main.py
@@ -8,6 +8,15 @@ import re
 import hmac
 from google.appengine.ext import db
 
+#import db models
+from dbmodel.user import User
+from dbmodel.plikes import Plikes
+from dbmodel.blogpost import BlogPost
+from dbmodel.bcomment import Bcomments
+
+#import handlers
+from handlers import Singup, Welcome, Login, Logout, NewPost
+
 # load template path folder
 template_dir = os.path.join(os.path.dirname(__file__), 'temp')
 # load templateing engine
@@ -19,36 +28,6 @@ USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
 PASS_RE = re.compile(r"^.{3,20}$")
 EMAIL_RE = re.compile(r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)")
 
-#data modles
-class User(db.Model):
-    """ application user model """
-    username = db.StringProperty(required=True)
-    password = db.StringProperty(required=True)
-    secret = db.StringProperty(required=True)
-    email = db.StringProperty()
-    created = db.DateTimeProperty(auto_now_add=True)
-
-class BlogPost(db.Model):
-    """ application blog model """
-    subject = db.StringProperty(required=True)
-    content = db.TextProperty(required=True)
-    created_by = db.StringProperty(required=True)
-    username = db.StringProperty(required=True)
-    created = db.DateTimeProperty(auto_now_add=True)
-    
-class Plikes(db.Model):
-    """ Blog like model"""
-    blog_id = db.StringProperty(required=True)
-    created_by = db.StringProperty(required=True)
-    created = db.DateTimeProperty(auto_now_add=True)
-
-class Bcomments(db.Model):
-    """ Blog comment model """
-    comment = db.TextProperty(required=True)
-    created_by = db.StringProperty(required=True)
-    blog_id = db.StringProperty(required=True)
-    username = db.StringProperty(required=True)
-    created = db.DateTimeProperty(auto_now_add=True)
 
 class Handler(webapp2.RequestHandler):
     """ general template render functions """
@@ -103,187 +82,187 @@ class MainPage(Handler):
         self.render("index.html", **index_dic)
 
 
-class Signup(Handler):
-    """ sing up processing class """
+# class Signup(Handler):
+#     """ sing up processing class """
 
-    def val_pass(self, password, verify):
-        """ check and verify password """
-        if PASS_RE.match(password):
-            if password == verify:
-                return True, password
-            else:
-                return False, "Passwords Don't Match"
-        else:
-            return False, "Invalid Password"
+#     def val_pass(self, password, verify):
+#         """ check and verify password """
+#         if PASS_RE.match(password):
+#             if password == verify:
+#                 return True, password
+#             else:
+#                 return False, "Passwords Don't Match"
+#         else:
+#             return False, "Invalid Password"
 
-    def val_username(self, username):
-        """ username validation"""
-        if USER_RE.match(username):
-            return True, username
-        else:
-            return False, "Invalid Username"
+#     def val_username(self, username):
+#         """ username validation"""
+#         if USER_RE.match(username):
+#             return True, username
+#         else:
+#             return False, "Invalid Username"
 
-    def val_email(self, email):
-        """ validate email address """
-        if email <> "":
-            if EMAIL_RE.match(email):
-                return True, email
-            return False, "Invalid Email"
-        else:
-            return True, "No Email"
+#     def val_email(self, email):
+#         """ validate email address """
+#         if email <> "":
+#             if EMAIL_RE.match(email):
+#                 return True, email
+#             return False, "Invalid Email"
+#         else:
+#             return True, "No Email"
 
-    def is_user(self, username):
-        """ checks if username exist"""
-        row = db.GqlQuery("select * from User where username = :1", username)
-        if row.fetch(limit=1) == []:
-            return True
-        else:
-            return False
+#     def is_user(self, username):
+#         """ checks if username exist"""
+#         row = db.GqlQuery("select * from User where username = :1", username)
+#         if row.fetch(limit=1) == []:
+#             return True
+#         else:
+#             return False
 
-    def get(self):
-        """ render Signup page"""
-        signup = {'title': "Subscribe to my Blogs",
-                  "username": self.request.get("username")}
-        self.render("signup.html", **signup)
+#     def get(self):
+#         """ render Signup page"""
+#         signup = {'title': "Subscribe to my Blogs",
+#                   "username": self.request.get("username")}
+#         self.render("signup.html", **signup)
 
-    def post(self):
-        """ validate and create user"""
-        user_val, user = self.val_username(self.request.get("username"))
-        pass_val, password = self.val_pass(
-            self.request.get("password"), self.request.get("verify"))
-        email = self.request.get("email")
-        is_user = self.is_user(user)
-        #ramdon generated secret
-        secret = ''.join(random.choice(
-            string.ascii_uppercase + string.digits) for _ in range(8))
-        if email is not None or email <> "":
-            email_val, email = self.val_email(email)
-        else:
-            email_val = True
-        if user_val and pass_val and email_val and is_user:
-            hash_pass = self.hash_str(password, secret)
-            user = User(username=user, password=hash_pass,
-                        secret=secret, email=email)
-            user.put()
-            user_id = user.key().id()
-            self.response.headers.add_header(
-                'Set-Cookie', 'user_id=%s|%s; Path=/' % (user_id, hash_pass))
-            self.redirect("/welcome")
-        else:
-            userinv = ""
-            passinv = ""
-            emailinv = ""
-            if user_val == False:
-                userinv = "Invalid Username"
-            if pass_val == False:
-                passinv = "Invalid Password"
-            if email_val == False:
-                emailinv = "Invalid Email"
-            if is_user == False:
-                userinv = "User Already exist"
-            error_dic = {"useinv": userinv, "passinv": passinv,
-                         "emailinv": emailinv, "username": user_val, "email": email}
-            self.render("signup.html", **error_dic)
-
-
-class Welcome(Handler):
-    """ main user landing page after login """
-
-    def post_by_user(self, created_by):
-        """ get post by user """
-        records = db.GqlQuery(
-            "select * from BlogPost where created_by = :1 order by created desc", created_by)
-        return records.fetch(limit=10)
-
-    def get(self):
-        """ render single user post"""
-        time.sleep(1) #added sleep to allow db to update
-        user_d = self.request.cookies.get('user_id')
-        if user_d:
-            user_id, user_hash = user_d.split("|")
-            user = User.get_by_id(int(user_id))
-            if user:
-                if user.password == user_hash:
-                    blog_post = self.post_by_user(user_id)
-                    details = {"username": str(user.username).upper(
-                    ), "logout": "Logout", "blog_post": blog_post}
-                    self.render("welcome.html", **details)
-                else:
-                    self.redirect("/signup")
-            else:
-                self.redirect("/signup")
-        else:
-            self.redirect("/signup")
+#     def post(self):
+#         """ validate and create user"""
+#         user_val, user = self.val_username(self.request.get("username"))
+#         pass_val, password = self.val_pass(
+#             self.request.get("password"), self.request.get("verify"))
+#         email = self.request.get("email")
+#         is_user = self.is_user(user)
+#         # ramdon generated secret
+#         secret = ''.join(random.choice(
+#             string.ascii_uppercase + string.digits) for _ in range(8))
+#         if email is not None or email <> "":
+#             email_val, email = self.val_email(email)
+#         else:
+#             email_val = True
+#         if user_val and pass_val and email_val and is_user:
+#             hash_pass = self.hash_str(password, secret)
+#             user = User(username=user, password=hash_pass,
+#                         secret=secret, email=email)
+#             user.put()
+#             user_id = user.key().id()
+#             self.response.headers.add_header(
+#                 'Set-Cookie', 'user_id=%s|%s; Path=/' % (user_id, hash_pass))
+#             self.redirect("/welcome")
+#         else:
+#             userinv = ""
+#             passinv = ""
+#             emailinv = ""
+#             if user_val == False:
+#                 userinv = "Invalid Username"
+#             if pass_val == False:
+#                 passinv = "Invalid Password"
+#             if email_val == False:
+#                 emailinv = "Invalid Email"
+#             if is_user == False:
+#                 userinv = "User Already exist"
+#             error_dic = {"useinv": userinv, "passinv": passinv,
+#                          "emailinv": emailinv, "username": user_val, "email": email}
+#             self.render("signup.html", **error_dic)
 
 
-class Login(Handler):
-    """ application login """
+# class Welcome(Handler):
+#     """ main user landing page after login """
 
-    def get(self):
-        """ render login page """
-        self.render("login.html")
+#     def post_by_user(self, created_by):
+#         """ get post by user """
+#         records = db.GqlQuery(
+#             "select * from BlogPost where created_by = :1 order by created desc", created_by)
+#         return records.fetch(limit=10)
 
-    def post(self):
-        """ process user login request """
-        user = self.request.get("username")
-        password = self.request.get("password")
-        error = ""
-        u_query = db.GqlQuery("select * from User where username = :1", user)
-        user_d = u_query.fetch(limit=1)
-        if user_d <> []:
-            secret = str(user_d[0].secret)
-            pass_hash = self.hash_str(password, secret) #create hash
-            if user_d[0].password == pass_hash: #validate hashed password
-                user_id = user_d[0].key().id()
-                self.response.headers.add_header(
-                    'Set-Cookie', 'user_id=%s|%s; Path=/ ' % (user_id, pass_hash))
-                self.redirect("/")
-            else:
-                error = "Invalid Password"
-                self.render("login.html", **{"username": user, "error": error})
-        else:
-            error = "no such user !!"
-            self.render("login.html", **{"username": user, "error": error})
-
-
-class Logout(Handler):
-
-    def get(self):
-        """ Delete user cookie """
-        self.response.delete_cookie('user_id')
-        self.redirect("/login")
+#     def get(self):
+#         """ render single user post"""
+#         time.sleep(1)  # added sleep to allow db to update
+#         user_d = self.request.cookies.get('user_id')
+#         if user_d:
+#             user_id, user_hash = user_d.split("|")
+#             user = User.get_by_id(int(user_id))
+#             if user:
+#                 if user.password == user_hash:
+#                     blog_post = self.post_by_user(user_id)
+#                     details = {"username": str(user.username).upper(
+#                     ), "logout": "Logout", "blog_post": blog_post}
+#                     self.render("welcome.html", **details)
+#                 else:
+#                     self.redirect("/signup")
+#             else:
+#                 self.redirect("/signup")
+#         else:
+#             self.redirect("/signup")
 
 
-class NewPost(Handler):
-    """ new blog post class """
+# class Login(Handler):
+#     """ application login """
 
-    def get(self):
-        """ render new blog post form """
-        user_d = self.request.cookies.get('user_id')
-        if user_d:
-            details = {"logout": "Logout"}
-            self.render("new.html", **details)
-        else:
-            self.redirect("/login")
+#     def get(self):
+#         """ render login page """
+#         self.render("login.html")
 
-    def post(self):
-        """ create new blog post """
-        subject = self.request.get("subject")
-        post = self.request.get("post")
-        user_d = self.request.cookies.get('user_id').split("|")[0]
-        if user_d:
-            if subject and post:
-                user = User.get_by_id(int(user_d))
-                new_post = BlogPost(
-                    subject=subject, content=post, created_by=user_d, username=user.username)
-                new_post.put()
-                new_post_id = new_post.key().id()
-                self.redirect("/single?post_id=" + str(new_post_id))
-            else:
-                details = {"error": "you need both subject and content"}
-                self.render("new.html", **details)
-        else:
-            self.redirect("/login")
+#     def post(self):
+#         """ process user login request """
+#         user = self.request.get("username")
+#         password = self.request.get("password")
+#         error = ""
+#         u_query = db.GqlQuery("select * from User where username = :1", user)
+#         user_d = u_query.fetch(limit=1)
+#         if user_d <> []:
+#             secret = str(user_d[0].secret)
+#             pass_hash = self.hash_str(password, secret)  # create hash
+#             if user_d[0].password == pass_hash:  # validate hashed password
+#                 user_id = user_d[0].key().id()
+#                 self.response.headers.add_header(
+#                     'Set-Cookie', 'user_id=%s|%s; Path=/ ' % (user_id, pass_hash))
+#                 self.redirect("/")
+#             else:
+#                 error = "Invalid Password"
+#                 self.render("login.html", **{"username": user, "error": error})
+#         else:
+#             error = "no such user !!"
+#             self.render("login.html", **{"username": user, "error": error})
+
+
+# class Logout(Handler):
+
+#     def get(self):
+#         """ Delete user cookie """
+#         self.response.delete_cookie('user_id')
+#         self.redirect("/login")
+
+
+# class NewPost(Handler):
+#     """ new blog post class """
+
+#     def get(self):
+#         """ render new blog post form """
+#         user_d = self.request.cookies.get('user_id')
+#         if user_d:
+#             details = {"logout": "Logout"}
+#             self.render("new.html", **details)
+#         else:
+#             self.redirect("/login")
+
+#     def post(self):
+#         """ create new blog post """
+#         subject = self.request.get("subject")
+#         post = self.request.get("post")
+#         user_d = self.request.cookies.get('user_id').split("|")[0]
+#         if user_d:
+#             if subject and post:
+#                 user = User.get_by_id(int(user_d))
+#                 new_post = BlogPost(
+#                     subject=subject, content=post, created_by=user_d, username=user.username)
+#                 new_post.put()
+#                 new_post_id = new_post.key().id()
+#                 self.redirect("/single?post_id=" + str(new_post_id))
+#             else:
+#                 details = {"error": "you need both subject and content"}
+#                 self.render("new.html", **details)
+#         else:
+#             self.redirect("/login")
 
 
 class SinglePost(Handler):
@@ -299,10 +278,13 @@ class SinglePost(Handler):
         """ render single blog post """
         post_id = self.request.get("post_id")
         blogpost = BlogPost.get_by_id(int(post_id))
-        comments = self.get_comments(post_id)
-        details = {"blogpost": blogpost, "logout": 'Logout', 'post_id': post_id,
-                   'comments': comments}
-        self.render("/single.html", **details)
+        if blogpost:
+            comments = self.get_comments(post_id)
+            details = {"blogpost": blogpost, "logout": 'Logout',
+                       'post_id': post_id, 'comments': comments}
+            self.render("/single.html", **details)
+        else:
+            self.redirect('/')
 
 
 class Edit(Handler):
@@ -461,7 +443,7 @@ class EditComment(Handler):
         if user_d:
             comment = Bcomments.get_by_id(int(com_id))
             if comment.created_by == user_d:
-                context = {"logout": 'Logout', 'comment':comment}
+                context = {"logout": 'Logout', 'comment': comment}
                 self.render('/edit_comment.html', **context)
             else:
                 self.redirect(
@@ -526,7 +508,7 @@ class ErrorHandler(Handler):
 
 
 app = webapp2.WSGIApplication([
-    ('/', MainPage), ('/signup', Signup), ('/welcome', Welcome),
+    ('/', MainPage), ('/signup', Singup), ('/welcome', Welcome),
     ('/login', Login), ('/logout', Logout), ('/new', NewPost),
     ('/single', SinglePost), ('/edit', Edit), ('/delete', Delete),
     ('/like', Likes), ('/error', ErrorHandler), ('/comment', Comment),
